# Potential escape optimization