Opening as a draft, to gauge your interest first.

One part of the PR is a fairly straightforward optimization, if we assume most strings don't need escaping we can save lots of work by optimistically scanning it it first.

I think this assumption is fair, given how ERB is used nowadays, where all interpolated strings are escaped, so aside from rare cases, only a minority need to be escaped.

If we take this assumption further, having to dup the string even when it's already safe is wasteful. So another API (currently `html_escape!`, but any name would do) which return the same string if escaping wasn't needed.

Given in the context of ERB interpolation, the code end up as `buffer << ERB::Util.escape(title)`, there's no need to dup to avoid mutation.

@k0kubun any opinions? Do you want me to cleanup either or both of these changes?

```
ruby 3.4.2 (2025-02-15 revision d2930f8e7a) +PRISM [arm64-darwin24]
Warming up --------------------------------------
              before     1.364M i/100ms
               after     1.909M i/100ms
                bang     2.829M i/100ms
Calculating -------------------------------------
              before     13.660M (± 0.3%) i/s   (73.21 ns/i) -     69.550M in   5.091551s
               after     19.157M (± 0.4%) i/s   (52.20 ns/i) -     97.340M in   5.081359s
                bang     28.330M (± 0.6%) i/s   (35.30 ns/i) -    144.274M in   5.092794s

Comparison:
              before: 13660076.4 i/s
                bang: 28329932.0 i/s - 2.07x  faster
               after: 19156538.1 i/s - 1.40x  faster

ruby 3.4.2 (2025-02-15 revision d2930f8e7a) +PRISM [arm64-darwin24]
Warming up --------------------------------------
              before     1.138M i/100ms
               after     1.400M i/100ms
                bang     1.411M i/100ms
Calculating -------------------------------------
              before     11.367M (± 4.0%) i/s   (87.97 ns/i) -     56.900M in   5.016941s
               after     13.935M (± 0.6%) i/s   (71.76 ns/i) -     69.980M in   5.022190s
                bang     13.884M (± 0.4%) i/s   (72.02 ns/i) -     70.552M in   5.081496s

Comparison:
              before: 11366937.1 i/s
               after: 13934530.2 i/s - 1.23x  faster
                bang: 13884325.4 i/s - 1.22x  faster
```

```
require 'bundler/inline'
gemfile do
  gem 'erb', path: "."
  gem 'benchmark-ips'
end

require 'benchmark/ips'
require 'erb'
require 'erb/escape'

class << ERB::Util
  def html_escape_old(s)
    CGI.escapeHTML(s.to_s)
  end
end

Benchmark.ips do |x|
  s = 'hello world'
  x.report('before') { ERB::Util.html_escape_old(s) }
  x.report('after')  { ERB::Util.html_escape(s) }
  x.report('bang')  { ERB::Util.html_escape!(s) }
  x.compare!(order: :baseline)
end

Benchmark.ips do |x|
  s = 'hello wo<ld'
  x.report('before') { ERB::Util.html_escape_old(s) }
  x.report('after')  { ERB::Util.html_escape(s) }
  x.report('bang')  { ERB::Util.html_escape!(s) }
  x.compare!(order: :baseline)
end
```