@k0kubun 2025-08-07
> gauge your interest
>
> having to dup the string even when it's already safe is wasteful. So another API (currently html_escape!, but any name would do) which return the same string if escaping wasn't needed.

Feel free to replace the current implementation with your `optimized_escape_html` without adding `html_escape!`. `old_optimized_escape_html` does not dup a given argument on the no-escape case. It's part of the reason why `ERB::Escape.html_escape` exists separately from `CGI.escapeHTML`. 

The speedup of `bang` on the no-escape case seems to come from the fact that it does not create a temporary buffer and copy the content to it, not from avoiding the object copy. It's interesting that `after` is faster than `before` but I guess `rb_str_dup` is faster than writing to the temporary buffer.

`before` vs `after` on the single-escape case is also interesting, but it might be because the first chunk becomes `memcpy` for `after`? It could be verified by benchmarking them with `'<ello world'` and `'h<llo world'`.