@byroot 2025-08-07
> `old_optimized_escape_html` does not dup a given argument on the no-escape case.

Oh shit 🤦 . So I can combine both. That's perfect.